feat(main.go): Splitted main.go into 4 files
